feat: enhance product and order management with new resources and relationships
- Added `BranchProductResource` and `OrderResource` for managing branch-specific products and orders.
- Implemented `ListBranchProducts` and `ViewOrder` pages for displaying product and order details.
- Updated `Branch` and `Product` models to include relationships for managing products and orders effectively.
- Enhanced `ProductsTable` and `OrdersTable` with new columns and filters for better data management.
- Introduced pivot attributes for products in branches, allowing for price and availability management.
- Improved serialization in `ProductResource` to include pricing details from the pivot table.

+++ database/migrations/2024_11_06_171802_create_orders_table.php
@@ -24,9 +24,9 @@ return new class extends Migration {
 
             $table->foreignId('branch_id')->constrained('branches');
 
-            $table->double('total_items_amount');
-            $table->double('delivery_amount')->default(0);
-            $table->double('total');
+            $table->decimal('total_items_amount', 10, 2);
+            $table->decimal('delivery_amount', 10, 2)->default(0);
+            $table->decimal('total', 10, 2);
 
             $table->text('notes')->nullable();
             $table->timestamps();
